refactor(helpers): Make one rule map the two URL namespaces
The same documentation page has two addresses - the file under
`/static/projects/` and the one vdoc's own router answers - and after the
previous commit two functions knew how to map between them:
`sanitizeDocuUri` stripped the prefix off the raw string with
`href.replace('static/projects/', '')`, and `toReadableHref` did it on the
parsed pathname.

`stripFramePrefix` is now the single rule, and both use it. Anchored at the
start of the path rather than replacing the prefix wherever it appears,
which also corrects a page whose own path spells the prefix out: for
`/static/projects/proj/1.0/static/projects/deep.html`, `sanitizeDocuUri`
used to report a `_splat` of `deep.html` and swallow the segment. Nothing
read that `_splat` outside its own tests, so this is a latent case being
closed rather than a live bug being fixed.

The two functions stay separate, because they answer to opposite
requirements and cannot be collapsed without breaking one of them:

- `sanitizeDocuUri` derives router parameters, so it normalizes. An empty
  fragment and a trailing slash are noise there and get dropped, and an
  address that is not a documentation page is an error rather than a value.
- `toReadableHref` writes into the `href` of a link inside a framed
  document, so it must preserve the address exactly - the empty fragment of
  `<a href="#">` included, which `tests/ui/testLinkSubstitution.spec.ts`
  asserts - and hand back anything it does not recognise untouched rather
  than throw.

A doc comment on `sanitizeDocuUri` now records that, so the question does
not have to be re-derived from the two implementations.

With that settled, the address helpers move to where the address knowledge
already lived: `RouteHelpers.ts` holds `FRAME_PATH_PREFIX`,
`stripFramePrefix`, `toReadableHref`, `toFrameHref` and `sanitizeDocuUri`,
while `IFrame.ts` keeps what reads and drives the framed document
(`parseIFrameHref`, `toggleDocumentationColorScheme`). The dependency runs
one way. Their unit tests move along with them.

Two new test groups: `stripFramePrefix` on its own, including the anchoring;
and one that pins `sanitizeDocuUri` and `toReadableHref` to naming the same
page for ordinary addresses, so the two rules cannot drift apart again
unnoticed.
