+++ src/ui/components/IFrame.tsx
@@ -8,8 +8,8 @@ import { useRouterState } from '@tanstack/react-router'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useIFrameScroll } from '../contexts/IFrameScrollContext'
 import { hookFramedDocument } from '../helpers/FramedDocument'
-import { parseIFrameHref, toFrameHref, toggleDocumentationColorScheme, toReadableHref } from '../helpers/IFrame'
-import { sanitizeDocuUri } from '../helpers/RouteHelpers'
+import { parseIFrameHref, toggleDocumentationColorScheme } from '../helpers/IFrame'
+import { sanitizeDocuUri, toFrameHref, toReadableHref } from '../helpers/RouteHelpers'
 import type { EffectiveColorMode } from '../interfacesAndTypes/ColorModes'
 import { testIDs } from '../interfacesAndTypes/testIDs'
