+++ src/ui/helpers/IFrame.ts
@@ -1,5 +1,6 @@
 import type { RefObject } from 'react'
 import type { EffectiveColorMode } from '../interfacesAndTypes/ColorModes'
+import { FRAME_PATH_PREFIX } from './RouteHelpers'
 
 export interface IFrameLocation {
   name: string
@@ -8,61 +9,6 @@ export interface IFrameLocation {
   search: URLSearchParams
   hash: string
   title: string
-}
-
-/**
- * Path prefix under which vdoc serves the published documentation files themselves.
- *
- * The files need a namespace of their own because relative links inside a framed page resolve
- * against the page's own address: without the prefix, a click inside the frame would load vdoc's
- * application into the frame instead of the next documentation page.
- */
-export const FRAME_PATH_PREFIX = '/static/projects/'
-
-/**
- * The readable form of `href`: the address vdoc's own router answers for the same page.
- *
- * The same page has two addresses - the file under {@link FRAME_PATH_PREFIX} and vdoc's own
- * `/{project}/{version}/...` - and mapping between them is a pure function in both directions,
- * with {@link toFrameHref} as the inverse. Nothing has to be remembered per link.
- *
- * Foreign addresses are returned verbatim, addresses of vdoc's own origin absolute, and applying
- * this to an address that is already readable changes nothing.
- *
- * @param href The address to convert, absolute or relative to `origin`.
- * @param origin The origin vdoc is served from.
- */
-export function toReadableHref(href: string, origin: string = window.location.origin): string {
-  const url = new URL(href, origin)
-  if (url.origin !== origin) {
-    return href
-  }
-  if (url.pathname.startsWith(FRAME_PATH_PREFIX)) {
-    url.pathname = `/${url.pathname.slice(FRAME_PATH_PREFIX.length)}`
-  }
-  return url.href
-}
-
-/**
- * The frame form of `href`: the address that reaches the documentation file itself.
- *
- * The inverse of {@link toReadableHref}, and idempotent for the same reason: this runs while a
- * click is being handled, and what the anchor carries at that moment is not certain. A framework
- * that re-renders a link restores the address the documentation authored, so the same link can
- * arrive in either form and must end up at the same file either way.
- *
- * @param href The address to convert, absolute or relative to `origin`.
- * @param origin The origin vdoc is served from.
- */
-export function toFrameHref(href: string, origin: string = window.location.origin): string {
-  const url = new URL(href, origin)
-  if (url.origin !== origin) {
-    return href
-  }
-  if (!url.pathname.startsWith(FRAME_PATH_PREFIX)) {
-    url.pathname = `${FRAME_PATH_PREFIX}${url.pathname.replace(/^\//, '')}`
-  }
-  return url.href
 }
 
 export function toggleDocumentationColorScheme(
